Rename home --> dashboard.
