+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/woodsmoke"
-    tools:context="com.eskaya.movie_application.presentation.home.HomeFragment">
+    tools:context="com.eskaya.movie_application.presentation.dashboard.DashboardFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
